clear font settings and enforce setting a base
When you did not set a `baseStyle`, previously styled text would not
revert to the text view's default font. Removing font settings falls
back to system defaults, though. So we need to make sure that users
keep this in mind and enforce setting a baseStyle that supersedes
`textView.font` and similar settings.